Fix AgileAggs bug (#192)
We were doing an explicitly unsafe conversion from []byte to string to
avoid copying, but this caused some memory issues. In particular, the
`combiner` map that it had a certain string as a key when in fact it did
not, because after we added that string as a key we changed the string's
data; the map apparently does not handle this, as strings in Go are
supposed to immutable.